✨헤어샵 본인인증로직 추가

+++ src/main/java/com/prgms/kokoahairshop/hairshop/controller/HairshopController.java
@@ -3,12 +3,16 @@ package com.prgms.kokoahairshop.hairshop.controller;
 import com.prgms.kokoahairshop.hairshop.dto.CreateHairshopRequest;
 import com.prgms.kokoahairshop.hairshop.dto.HairshopResponse;
 import com.prgms.kokoahairshop.hairshop.dto.ModifyHairshopRequest;
+import com.prgms.kokoahairshop.hairshop.entity.Hairshop;
 import com.prgms.kokoahairshop.hairshop.service.HairshopService;
+import com.prgms.kokoahairshop.user.entity.User;
 import java.net.URI;
 import javax.validation.Valid;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -18,6 +22,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@Slf4j
 @RestController
 @RequestMapping("/hairshops")
 public class HairshopController {
@@ -30,9 +35,12 @@ public class HairshopController {
 
     @PostMapping
     public ResponseEntity<Long> insert(
-        @Valid @RequestBody CreateHairshopRequest createHairshopRequest) {
+        @Valid @RequestBody CreateHairshopRequest createHairshopRequest, @AuthenticationPrincipal User user) {
+        if (!user.getId().equals(createHairshopRequest.getUserId())){
+            throw new IllegalArgumentException("본인의 헤어샵만 생성할 수 있습니다.");
+        }
         HairshopResponse insert = hairshopService.insert(createHairshopRequest);
-        return ResponseEntity.created(URI.create("/api/v1/hairshops/" + insert.getId()))
+        return ResponseEntity.created(URI.create("/hairshops/" + insert.getId()))
             .body(insert.getId());
     }
 
@@ -49,13 +57,21 @@ public class HairshopController {
 
     @PatchMapping
     public ResponseEntity<Object> modify(
-        @Valid @RequestBody ModifyHairshopRequest modifyHairshopRequest) {
+        @Valid @RequestBody ModifyHairshopRequest modifyHairshopRequest, @AuthenticationPrincipal User user) {
+        if (!user.getId().equals(modifyHairshopRequest.getUserId())){
+            throw new IllegalArgumentException("본인의 헤어샵만 수정할 수 있습니다.");
+        }
         hairshopService.update(modifyHairshopRequest);
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deleteById(@PathVariable Long id) {
+    public ResponseEntity<Object> deleteById(@PathVariable Long id, @AuthenticationPrincipal User user) {
+        Hairshop hairshop = hairshopService.findHairshopById(id);
+        log.info("hairshop: {}",hairshop);
+        if (!user.getId().equals(hairshop.getUser().getId())){
+            throw new IllegalArgumentException("본인의 헤어샵만 삭제할 수 있습니다.");
+        }
         hairshopService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
